Replaced int with size_t in RectangularDetector constructor

+++ Core/Instrument/RectangularDetector.cpp
@@ -21,7 +21,7 @@
 #include "MathConstants.h"
 #include "Units.h"
 
-RectangularDetector::RectangularDetector(int nxbins, double width, int nybins, double height)
+RectangularDetector::RectangularDetector(size_t nxbins, double width, size_t nybins, double height)
     : m_u0(0.0)
     , m_v0(0.0)
     , m_direction(kvector_t(0.0, -1.0, 0.0))
